adaptive start multiple workflows
The script should only start a batch of workflows if there are less active workflows as an threshold value. This should guarantee that the system will be responding.
The algorithmus get the active workflows count from the server before running each iteration of the start-workflow script. If the value is greather than `RUN_WORKFLOWS_CONCURENTLY`, then it waits until the value is dropped below `START_WORKFLOWS_THRESHOLD`.