Removed group types
The group types (closed or open) are not used in the script anymore.
The were formerly used to allow certain members to be added without asking the user again.
Now, as user permissions can be stored, this is not neccesary anymore.
The user is asked for every member, if no permissions has been specified before.