Hide bot PRs in changelog
When looking through the changelog in the docs to see what has
changed, most people will not care about the semi-automatic
dependency updates. Instead, they pose a hindrance as they need to
be mentally filtered out.

This PR changes the behaviour of the changelog generation, filtering
bot pull requests to the bottom and hiding them in a html dropdown.